Move `codeformat` folder to tools
