+++ tools/code-batch-process.go
@@ -16,7 +16,7 @@ import (
 	"strconv"
 	"strings"
 
-	"code.gitea.io/gitea/build/codeformat"
+	"code.gitea.io/gitea/tools/codeformat"
 )
 
 // Windows has a limitation for command line arguments, the size can not exceed 32KB.
